requirements.txt
tensorflow==2.15.0
numpy==1.24.3
Pillow==10.0.0
scikit-learn==1.3.0
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
openai==1.3.7
python-dotenv==1.0.0 